+++ private/WebInspector/RWIProtocolNetworkResponse.h
@@ -20,9 +20,9 @@
 @property(copy, nonatomic) NSString *headersText;
 @property(retain, nonatomic) RWIProtocolNetworkHeaders *headers;
 @property(copy, nonatomic) NSString *statusText;
-@property(nonatomic) double status;
+@property(nonatomic) int status;
 @property(copy, nonatomic) NSString *url;
-- (id)initWithUrl:(id)arg1 status:(double)arg2 statusText:(id)arg3 headers:(id)arg4 mimeType:(id)arg5 source:(long long)arg6;
+- (id)initWithUrl:(id)arg1 status:(int)arg2 statusText:(id)arg3 headers:(id)arg4 mimeType:(id)arg5 source:(long long)arg6;
 
 @end
