+++ private/iWorkImport/TPFootnotePropertyStrings.h
@@ -11,13 +11,13 @@ __attribute__((visibility("hidden")))
 {
 }
 
-+ (id)stringForFootnoteNumbering:(int)arg1;
-+ (id)stringForFootnoteFormat:(int)arg1;
-+ (id)stringForFootnoteKind:(int)arg1;
++ (id)stringForFootnoteNumbering:(long long)arg1;
++ (id)stringForFootnoteFormat:(long long)arg1;
++ (id)stringForFootnoteKind:(long long)arg1;
 + (id)titleForFootnoteNumbering;
 + (id)titleForFootnoteFormat;
 + (id)titleForFootnoteKind;
-+ (id)insertStringForFootnoteKind:(int)arg1;
++ (id)insertStringForFootnoteKind:(long long)arg1;
 
 @end
