+++ private/iWorkImport/TPPageLayoutInfoProvider-Protocol.h
@@ -12,7 +12,7 @@
 - (NSString *)displayPageNumberForPageIndex:(unsigned long long)arg1;
 - (NSString *)displayPageNumberForCharIndex:(unsigned long long)arg1;
 - (_Bool)shouldHeaderFooterBeVisibleForPageIndex:(unsigned long long)arg1;
-- (int)contentFlagsForPageIndex:(unsigned long long)arg1;
+- (unsigned long long)contentFlagsForPageIndex:(unsigned long long)arg1;
 - (NSString *)pageNumberFormatForSectionOnPageIndex:(unsigned long long)arg1;
 - (unsigned long long)pageCountForPageIndex:(unsigned long long)arg1;
 - (unsigned long long)pageNumberForPageIndex:(unsigned long long)arg1;
