+++ private/iWorkImport/TPSectionHint.h
@@ -23,8 +23,8 @@ __attribute__((visibility("hidden")))
 - (id)initWithArchive:(const struct SectionHintArchive *)arg1 unarchiver:(id)arg2;
 - (void)setPageHints:(id)arg1;
 - (_Bool)containsDocumentPageIndex:(unsigned long long)arg1;
-- (_Bool)hasPageHintOfKind:(int)arg1 beforePageIndex:(unsigned long long)arg2;
-- (_Bool)hasPageHintOfKind:(int)arg1 atPageIndex:(unsigned long long)arg2;
+- (_Bool)hasPageHintOfKind:(long long)arg1 beforePageIndex:(unsigned long long)arg2;
+- (_Bool)hasPageHintOfKind:(long long)arg1 atPageIndex:(unsigned long long)arg2;
 - (void)trimPageHintsFromPageIndex:(unsigned long long)arg1;
 - (id)pageHintForPageIndex:(unsigned long long)arg1;
 @property(readonly, nonatomic) struct _NSRange documentPageRange;
